🐛fix(#132): 회원가입 생일 등록란 삭제
